+++ lib/patient_http/client.rb
@@ -12,6 +12,7 @@ module PatientHttp
         protocol: config.protocol
       )
       @response_reader = ResponseReader.new(@processor)
+      @request_preparer = RequestPreparer.new(config)
     end
 
     # Make an asynchronous HTTP request.
@@ -23,13 +24,16 @@ module PatientHttp
       async_response = nil
 
       begin
-        headers = request_headers(request, request_id)
-        url = config.secret_manager.resolve_url(request.url, request.secret_params)
+        outgoing = @request_preparer.prepare(request, request_id)
+        url = outgoing.url
+        headers = outgoing.headers.to_h
         body = Protocol::HTTP::Body::Buffered.wrap([request.body.to_s]) if request.body
         timeout = request.timeout || config.request_timeout
 
         Async::Task.current.with_timeout(timeout) do
           async_response = @client_pool.request(request.http_method, url, headers, body)
+          # Note: headers that appear multiple times (e.g. set-cookie) are
+          # flattened to a single joined string value.
           headers_hash = async_response.headers.to_h.transform_values(&:to_s)
           body = @response_reader.read_body(async_response, headers_hash)
 
@@ -66,18 +70,11 @@ module PatientHttp
     def connection_error?(exception)
       case exception
       when Async::TimeoutError, Errno::ECONNRESET, Errno::ECONNABORTED, Errno::EPIPE,
-           Errno::ECONNREFUSED, Errno::EHOSTUNREACH, SocketError, IOError
+           Errno::ECONNREFUSED, Errno::EHOSTUNREACH, Errno::ETIMEDOUT, SocketError, IOError
         true
       else
         false
       end
-    end
-
-    def request_headers(request, request_id)
-      headers = config.secret_manager.resolve_headers(request.headers.to_h)
-      headers["x-request-id"] = request_id
-      headers["user-agent"] ||= config.user_agent if config.user_agent
-      headers
     end
   end
 end